% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{pin_qsave}
\title{Deprecated functions}
\usage{
pin_qsave(board, x, name, nthreads = NULL, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{pin_qwrite()} instead of \code{pin_qsave()}.
}
\keyword{internal}
