% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinsqs-package.R
\docType{package}
\name{pinsqs-package}
\alias{pinsqs}
\alias{pinsqs-package}
\title{pinsqs: `qs` Multithreading Support for `pins`}
\description{
Provides convenience functions to incorporate multithreaded reading and writing of `.qs` files from the `qs` package into the `pins` infrastructure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://adamoshen.github.io/pinsqs}
  \item \url{https://github.com/adamoshen/pinsqs}
  \item Report bugs at \url{https://github.com/adamoshen/pinsqs/issues}
}

}
\author{
\strong{Maintainer}: Adam Shen \email{adamshen1@gmail.com}

}
\keyword{internal}
