% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinsqs-package.R
\docType{package}
\name{pinsqs-package}
\alias{pinsqs}
\alias{pinsqs-package}
\title{pinsqs: `qs` Multithreading Support for `pins`}
\description{
Provides convenience functions to incorporate multithreaded reading and writing of `.qs` files from the `qs` package into the `pins` infrastructure.
}
\author{
\strong{Maintainer}: Adam Shen \email{adam.shen@carleton.ca}

}
\keyword{internal}
